ci: adjust ssh server

+++ pack/ssh-server/rootfs/entrypoint.sh
@@ -2,13 +2,18 @@
 
 set -e
 
-if [ -n "$SSH_PUBLIC_KEY" ]; then
-  if ! [ -f /root/.ssh/authorized_keys ]; then
-    mkdir -p /root/.ssh
-    echo "$SSH_PUBLIC_KEY" >/root/.ssh/authorized_keys
-    chmod 600 /root/.ssh/authorized_keys
-  else
-    echo "!!! Warning: /root/.ssh/authorized_keys already exists, ignore SSH_PUBLIC_KEY env !!!" >&2
+if [ -f /var/run/sshd-authorized-keys/authorized_keys ]; then
+  echo "AuthorizedKeysCommand /get-authorized-keys.sh" >> /etc/ssh/sshd_config
+  echo "AuthorizedKeysCommandUser root" >> /etc/ssh/sshd_config
+else
+  if [ -n "$SSH_PUBLIC_KEY" ]; then
+    if ! [ -f /root/.ssh/authorized_keys ]; then
+      mkdir -p /root/.ssh
+      echo "$SSH_PUBLIC_KEY" >/root/.ssh/authorized_keys
+      chmod 600 /root/.ssh/authorized_keys
+    else
+      echo "!!! Warning: /root/.ssh/authorized_keys already exists, ignore SSH_PUBLIC_KEY env !!!" >&2
+    fi
   fi
 fi
